% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_calc_agreement.R
\name{cmd_calc_agreement}
\alias{cmd_calc_agreement}
\title{Calculate Agreement Metrics for Variables Extracted by Annotators}
\usage{
cmd_calc_agreement(x, cols, consensus_threshold = 0.6)
}
\arguments{
\item{x}{A data.frame or data.table containing the extracted variables
with columns for 'pmid', 'annotator_id', and the specified variable columns.}

\item{cols}{A character vector specifying the columns to be used for variable
extraction.}

\item{consensus_threshold}{A numeric value indicating the threshold for consensus
rate. Defaults to 0.6.}
}
\value{
A data.table containing the agreement metrics, including Fleiss' Kappa,
consensus rates, and other summary statistics for each PMID.
}
\description{
This function calculates various agreement metrics, including Fleiss' Kappa,
for variables extracted by multiple annotators. It summarizes the results
and provides consensus rates and other relevant statistics.
}
\examples{
library(data.table)
variables <- data.table(
  pmid = c(28817241, 28817241, 28817241, 28817241, 28817241, 28817241,
           28817242, 28817242, 28817242, 28817242, 28817242, 28817242),
  annotator_id = c("OSM5611", "OSM5611", "OSM5611", "OSM5611", "OSM5611", "OSM5611",
                   "OSM5611", "OSM5611", "OSM5611", "OSM5611", "OSM5611", "OSM5611"),
  variable_name = c("Heart Failure", "Vitamin D Deficiency", "Age", "Gender", 
                    "Education", "Ethnicity", "Diabetes", "Hypertension", 
                    "Age", "Gender", "Education", "Ethnicity"),
  variable_type = c("Outcome", "Exposure", "Covariate", "Covariate", "Covariate", 
                    "Covariate", "Outcome", "Exposure", "Covariate", "Covariate", 
                    "Covariate", "Covariate")
)
cols <- c("variable_name", "variable_type")
result <- cmd_calc_agreement(variables, cols)
print(result)
}
