% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompts.R
\docType{data}
\name{cmd_prompts}
\alias{cmd_prompts}
\title{Command Prompts for ResearchTextProcessor Package}
\format{
An object of class \code{list} of length 5.
}
\usage{
cmd_prompts
}
\description{
This list contains descriptions of various tasks related to processing and classifying scientific texts,
extracting key variables, and generating structured outputs for epidemiological and clinical research data.
}
\keyword{datasets}
