% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comradeGPT-package.R
\docType{package}
\name{comradeGPT-package}
\alias{comradeGPT}
\alias{comradeGPT-package}
\title{comradeGPT:}
\description{

}
\author{
\strong{Maintainer}: Jason Timm \email{JaTimm@salud.unm.edu}

}
\keyword{internal}
